@@ -113,42 +114,42 @@
     "cuid": "2.1.8",
     "eventsource": "1.0.7",
     "html-minifier": "4.0.0",
-    "humanize-duration": "3.23.0",
+    "humanize-duration": "3.23.1",
     "istanbul-lib-coverage": "3.0.0",
     "istanbul-lib-instrument": "4.0.3",
     "istanbul-lib-report": "3.0.0",
     "istanbul-reports": "3.0.2",
     "node-notifier": "7.0.1",
-    "playwright-core": "1.0.2",
+    "parse5": "6.0.0",
+    "playwright-core": "1.2.0",
     "proper-lockfile": "4.1.1",
     "regenerator-runtime": "0.13.5",
-    "rollup": "2.11.2",
+    "rollup": "2.21.0",
     "rollup-plugin-node-builtins": "2.1.2",
     "rollup-plugin-node-globals": "1.4.0",
     "source-map": "0.7.3",
     "string-width": "4.2.0",
     "supports-color": "7.1.0",
     "systemjs": "6.1.9",
-    "terser": "4.7.0",
+    "terser": "4.8.0",
     "tree-kill": "1.2.2",
     "wrap-ansi": "7.0.0"
   },
   "devDependencies": {
-    "@babel/plugin-transform-react-jsx": "7.9.4",
-    "@babel/plugin-transform-typescript": "7.9.4",
-    "@jsenv/assert": "2.1.0",
+    "@babel/plugin-transform-react-jsx": "7.10.4",
+    "@babel/plugin-transform-typescript": "7.10.4",
+    "@jsenv/assert": "2.2.1",
     "@jsenv/codecov-upload": "3.4.1",
     "@jsenv/core": "13.1.4",
     "@jsenv/eslint-config": "12.4.1",
-    "@jsenv/git-hooks": "1.3.3",
     "@jsenv/github-release-package": "1.2.2",
     "@jsenv/package-publish": "1.5.1",
     "@jsenv/prettier-check-project": "5.6.0",
     "babel-eslint": "11.0.0-beta.0",
-    "eslint": "7.1.0",
-    "playwright-chromium": "1.0.2",
-    "playwright-firefox": "1.0.2",
-    "playwright-webkit": "1.0.2",
+    "eslint": "7.4.0",
+    "playwright-chromium": "1.2.0",
+    "playwright-firefox": "1.2.0",
+    "playwright-webkit": "1.2.0",
     "prettier": "2.0.5",
     "react": "16.13.1",
     "redux": "4.0.5"
