@@ -9,7 +9,7 @@ import {
   writeFile,
   comparePathnames,
 } from "@jsenv/util"
-import { writeSourceMappingURL } from "../../sourceMappingURLUtils.js"
+import { appendSourceMappingAsExternalUrl } from "../../sourceMappingURLUtils.js"
 import { fetchUrl } from "../../fetchUrl.js"
 import { validateResponseStatusIsOk } from "../../validateResponseStatusIsOk.js"
 import { transformJs } from "../../compiling/js-compilation-service/transformJs.js"
@@ -479,7 +479,7 @@ const transformAsyncInsertedByRollup = async ({
       })
 
       await Promise.all([
-        writeFile(bundleFileUrl, writeSourceMappingURL(code, `./${bundleFilename}.map`)),
+        writeFile(bundleFileUrl, appendSourceMappingAsExternalUrl(code, `./${bundleFilename}.map`)),
         writeFile(`${bundleFileUrl}.map`, JSON.stringify(map)),
       ])
     }),
