@@ -1,4 +1,5 @@
-import { urlToFileSystemPath, writeFile, readFileSystemNodeStat, bufferToEtag } from "@jsenv/util"
+import { urlToFileSystemPath, bufferToEtag } from "@jsenv/util"
+import { writeFileContent, testFilePresence } from "./fs-optimized-for-cache.js"
 import { resolveSourceFileUrl, resolveMetaJsonFileUrl, resolveAssetFileUrl } from "./locaters.js"
 
 export const updateMeta = async ({
@@ -15,31 +16,38 @@ export const updateMeta = async ({
   const { compiledSource, contentType, assets, assetsContent } = compileResult
   let { sources, sourcesContent } = compileResult
 
-  // ensure source that does not leads to concrete files are not capable to invalidate the cache
-  const sourceExists = await Promise.all(
-    sources.map(async (source) => {
-      const sourceFileUrl = resolveSourceFileUrl({ source, compiledFileUrl })
-      const sourceStats = await readFileSystemNodeStat(sourceFileUrl, { nullIfNotFound: true })
-      if (sourceStats === null) {
-        // this can lead to cache never invalidated by itself
-        // it's a very important warning
-        logger.warn(`a source file cannot be found ${sourceFileUrl}.
--> excluding it from meta.sources & meta.sourcesEtag`)
-        return false
-      }
-      return true
-    }),
-  )
-  sources = sources.filter((source, index) => sourceExists[index])
-  sourcesContent = sourcesContent.filter((sourceContent, index) => sourceExists[index])
-
   const promises = []
   if (isNew || isUpdated) {
+    // ensure source that does not leads to concrete files are not capable to invalidate the cache
+    const sourceExists = await Promise.all(
+      sources.map(async (source) => {
+        const sourceFileUrl = resolveSourceFileUrl({ source, compiledFileUrl })
+        const sourceFileExists = await testFilePresence(sourceFileUrl)
+        if (sourceFileExists) {
+          return true
+        }
+        // this can lead to cache never invalidated by itself
+        // it's a very important warning
+        logger.warn(`a source file cannot be found -> excluded from meta.sources & meta.sourcesEtag.
+--- source ---
+${source}
+-- source url ---
+${sourceFileUrl}`)
+        return false
+      }),
+    )
+    sources = sources.filter((source, index) => sourceExists[index])
+    sourcesContent = sourcesContent.filter((sourceContent, index) => sourceExists[index])
+
     const { writeCompiledSourceFile = true, writeAssetsFile = true } = compileResult
 
     if (writeCompiledSourceFile) {
       logger.debug(`write compiled file at ${urlToFileSystemPath(compiledFileUrl)}`)
-      promises.push(writeFile(compiledFileUrl, compiledSource))
+      promises.push(
+        writeFileContent(compiledFileUrl, compiledSource, {
+          fileLikelyNotFound: isNew,
+        }),
+      )
     }
 
     if (writeAssetsFile) {
@@ -50,7 +58,9 @@ export const updateMeta = async ({
             asset,
           })
           logger.debug(`write compiled file asset at ${urlToFileSystemPath(assetFileUrl)}`)
-          return writeFile(assetFileUrl, assetsContent[index])
+          return writeFileContent(assetFileUrl, assetsContent[index], {
+            fileLikelyNotFound: isNew,
+          })
         }),
       )
     }
@@ -111,7 +121,11 @@ export const updateMeta = async ({
     })
 
     logger.debug(`write compiled file meta at ${urlToFileSystemPath(metaJsonFileUrl)}`)
-    promises.push(writeFile(metaJsonFileUrl, JSON.stringify(latestMeta, null, "  ")))
+    promises.push(
+      writeFileContent(metaJsonFileUrl, JSON.stringify(latestMeta, null, "  "), {
+        fileLikelyNotFound: isNew,
+      }),
+    )
   }
 
   return Promise.all(promises)
