@@ -1,9 +1,10 @@
 import { extname, basename } from "path"
-import { resolveDirectoryUrl, urlToRelativeUrl } from "@jsenv/util"
+import { resolveDirectoryUrl, urlToRelativeUrl, resolveUrl, urlIsInsideOf } from "@jsenv/util"
+import { jsenvCoreDirectoryUrl } from "../jsenvCoreDirectoryUrl.js"
 import { COMPILE_ID_GLOBAL_BUNDLE_FILES, COMPILE_ID_COMMONJS_BUNDLE_FILES } from "../CONSTANTS.js"
 import { generateBundleUsingRollup } from "../bundling/generateBundleUsingRollup.js"
 import { bundleToCompilationResult } from "../bundling/bundleToCompilationResult.js"
-import { serveCompiledFile } from "./serveCompiledFile.js"
+import { compileFile } from "./compileFile.js"
 
 export const serveBundle = async ({
   cancellationToken,
@@ -72,7 +73,12 @@ export const serveBundle = async ({
     })
   }
 
-  return serveCompiledFile({
+  // might want to put this to false while working on jsenv
+  // to that cache gets verified
+  const isJenvInternalFile =
+    false && urlIsInsideOf(originalFileUrl, resolveUrl("./src/internal/", jsenvCoreDirectoryUrl))
+
+  return compileFile({
     logger,
     projectDirectoryUrl,
     originalFileUrl,
@@ -83,5 +89,7 @@ export const serveBundle = async ({
     projectFileRequestedCallback,
     compile,
     request,
+    compileCacheSourcesValidation: !isJenvInternalFile,
+    compileCacheAssetsValidation: !isJenvInternalFile,
   })
 }
