@@ -7,7 +7,7 @@ import { EXECUTE_TEST_PARAMS } from "../TEST_PARAMS.js"
 const testDirectoryUrl = resolveUrl("./", import.meta.url)
 const testDirectoryRelativeUrl = urlToRelativeUrl(testDirectoryUrl, jsenvCoreDirectoryUrl)
 const jsenvDirectoryRelativeUrl = `${testDirectoryRelativeUrl}.jsenv/`
-const fileRelativeUrl = `${testDirectoryRelativeUrl}file.js`
+const fileRelativeUrl = `${testDirectoryRelativeUrl}file.html`
 
 await Promise.all(
   [launchChromium, launchFirefox, launchWebkit].map(async (launchBrowser) => {
@@ -21,6 +21,12 @@ await Promise.all(
     })
     const expected = {
       status: "completed",
+      namespace: {
+        "./file.js": {
+          status: "completed",
+          namespace: {},
+        },
+      },
     }
     assert({ actual, expected })
   }),
