@@ -13,12 +13,10 @@ import { EXECUTE_TEST_PARAMS } from "../TEST_PARAMS.js"
 const testDirectoryUrl = resolveUrl("./", import.meta.url)
 const testDirectoryRelativeUrl = urlToRelativeUrl(testDirectoryUrl, jsenvCoreDirectoryUrl)
 const jsenvDirectoryRelativeUrl = `${testDirectoryRelativeUrl}.jsenv/`
+const htmlFileRelativeUrl = `${testDirectoryRelativeUrl}file.html`
 const fileRelativeUrl = `${testDirectoryRelativeUrl}file.js`
 const testPlan = {
-  [fileRelativeUrl]: {
-    node: {
-      launch: launchNode,
-    },
+  [htmlFileRelativeUrl]: {
     chromium: {
       launch: launchChromium,
     },
@@ -27,6 +25,11 @@ const testPlan = {
     },
     webkit: {
       launch: launchWebkit,
+    },
+  },
+  [fileRelativeUrl]: {
+    node: {
+      launch: launchNode,
     },
   },
 }
@@ -48,6 +51,41 @@ const expected = {
     endMs: actual.summary.endMs,
   },
   report: {
+    [htmlFileRelativeUrl]: {
+      chromium: {
+        status: "completed",
+        namespace: {
+          "./file.js": {
+            status: "completed",
+            namespace: { default: "browser" },
+          },
+        },
+        runtimeName: "chromium",
+        runtimeVersion: assert.any(String),
+      },
+      firefox: {
+        status: "completed",
+        namespace: {
+          "./file.js": {
+            status: "completed",
+            namespace: { default: "browser" },
+          },
+        },
+        runtimeName: "firefox",
+        runtimeVersion: assert.any(String),
+      },
+      webkit: {
+        status: "completed",
+        namespace: {
+          "./file.js": {
+            status: "completed",
+            namespace: { default: "browser" },
+          },
+        },
+        runtimeName: "webkit",
+        runtimeVersion: assert.any(String),
+      },
+    },
     [fileRelativeUrl]: {
       node: {
         status: "completed",
@@ -55,31 +93,7 @@ const expected = {
           default: "node",
         },
         runtimeName: "node",
-        runtimeVersion: actual.report[fileRelativeUrl].node.runtimeVersion,
-      },
-      chromium: {
-        status: "completed",
-        namespace: {
-          default: "browser",
-        },
-        runtimeName: "chromium",
-        runtimeVersion: actual.report[fileRelativeUrl].chromium.runtimeVersion,
-      },
-      firefox: {
-        status: "completed",
-        namespace: {
-          default: "browser",
-        },
-        runtimeName: "firefox",
-        runtimeVersion: actual.report[fileRelativeUrl].firefox.runtimeVersion,
-      },
-      webkit: {
-        status: "completed",
-        namespace: {
-          default: "browser",
-        },
-        runtimeName: "webkit",
-        runtimeVersion: actual.report[fileRelativeUrl].webkit.runtimeVersion,
+        runtimeVersion: assert.any(String),
       },
     },
   },
