@@ -5,17 +5,21 @@ import { executeTestPlan, launchNode, launchChromium } from "../../../index.js"
 import { EXECUTE_TEST_PARAMS } from "../TEST_PARAMS.js"
 
 const testDirectoryUrl = resolveDirectoryUrl("./", import.meta.url)
-const testDirectoryRelativePath = urlToRelativeUrl(testDirectoryUrl, jsenvCoreDirectoryUrl)
-const jsenvDirectoryRelativeUrl = `${testDirectoryRelativePath}.jsenv/`
+const testDirectoryRelativeUrl = urlToRelativeUrl(testDirectoryUrl, jsenvCoreDirectoryUrl)
+const jsenvDirectoryRelativeUrl = `${testDirectoryRelativeUrl}.jsenv/`
+const htmlFileRelativeUrl = `${testDirectoryRelativeUrl}import-throw.html`
+const fileRelativeUrl = `${testDirectoryRelativeUrl}import-throw.js`
 const { coverageMap: actual } = await executeTestPlan({
   ...EXECUTE_TEST_PARAMS,
   executionLogLevel: "off",
   jsenvDirectoryRelativeUrl,
   testPlan: {
-    [`${testDirectoryRelativePath}import-throw.js`]: {
+    [htmlFileRelativeUrl]: {
       chromium: {
         launch: launchChromium,
       },
+    },
+    [fileRelativeUrl]: {
       node: {
         launch: launchNode,
       },
@@ -23,12 +27,12 @@ const { coverageMap: actual } = await executeTestPlan({
   },
   coverage: true,
   coverageConfig: {
-    [`${testDirectoryRelativePath}throw.js`]: true,
+    [`${testDirectoryRelativeUrl}throw.js`]: true,
   },
 })
 const expected = {
-  [`${testDirectoryRelativePath}throw.js`]: {
-    ...actual[`${testDirectoryRelativePath}throw.js`],
+  [`${testDirectoryRelativeUrl}throw.js`]: {
+    ...actual[`${testDirectoryRelativeUrl}throw.js`],
     s: { 0: 2, 1: 2 },
   },
 }
