@@ -15,7 +15,7 @@ const testDirectoryUrl = resolveDirectoryUrl("./", import.meta.url)
 const testDirectoryRelativeUrl = urlToRelativeUrl(testDirectoryUrl, jsenvCoreDirectoryUrl)
 const testDirectoryname = basename(testDirectoryRelativeUrl)
 const jsenvDirectoryRelativeUrl = `${testDirectoryRelativeUrl}.jsenv`
-const filename = `${testDirectoryname}.js`
+const filename = `${testDirectoryname}.html`
 const fileRelativeUrl = `${testDirectoryRelativeUrl}${filename}`
 const { origin: compileServerOrigin, outDirectoryRelativeUrl } = await startCompileServer({
   ...START_COMPILE_SERVER_TEST_PARAMS,
@@ -34,12 +34,16 @@ await Promise.all(
           outDirectoryRelativeUrl,
           compileServerOrigin,
         }),
-      collectNamespace: true,
     })
     const expected = {
       status: "completed",
       namespace: {
-        default: 42,
+        "./dynamic-import.js": {
+          status: "completed",
+          namespace: {
+            default: 42,
+          },
+        },
       },
     }
     assert({ actual, expected })
