@@ -15,7 +15,7 @@ const testDirectoryUrl = resolveUrl("./", import.meta.url)
 const testDirectoryRelativeUrl = urlToRelativeUrl(testDirectoryUrl, jsenvCoreDirectoryUrl)
 const testDirectoryBasename = basename(testDirectoryRelativeUrl)
 const jsenvDirectoryRelativeUrl = `${testDirectoryRelativeUrl}.jsenv`
-const filename = `${testDirectoryBasename}.js`
+const filename = `${testDirectoryBasename}.html`
 const fileRelativeUrl = `${testDirectoryRelativeUrl}${filename}`
 const { origin: compileServerOrigin, outDirectoryRelativeUrl } = await startCompileServer({
   ...START_COMPILE_SERVER_TEST_PARAMS,
@@ -37,7 +37,12 @@ const actual = await launchAndExecute({
 const expected = {
   status: "completed",
   namespace: {
-    default: 42,
+    "./generator.js": {
+      status: "completed",
+      namespace: {
+        default: 42,
+      },
+    },
   },
 }
 assert({ actual, expected })
