@@ -33,9 +33,9 @@ const actual = await launchAndExecute({
       outDirectoryRelativeUrl,
       compileServerOrigin,
     }),
-  collectNamespace: false,
 })
 const expected = {
   status: "completed",
+  namespace: {},
 }
 assert({ actual, expected })
